+++ lib/modules/navigation/navigation_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: navigation
-// Updated at: 2026-08-13 16:35:24
-// Commit: feat(navigation): add bi-directional event bus handler for navigation
+// Updated at: 2026-08-14 11:32:43
+// Commit: feat(navigation): add async navigation stream provider
 
 class navigationComponentState {
-  final String commitRef = '3581';
-  final String lastMessage = 'feat(navigation): add bi-directional event bus handler for navigation';
-  final DateTime timestamp = DateTime.parse('2026-08-13T16:35:24+05:00');
+  final String commitRef = '3604';
+  final String lastMessage = 'feat(navigation): add async navigation stream provider';
+  final DateTime timestamp = DateTime.parse('2026-08-14T11:32:43+05:00');
 }